Backend/requirements.txt
flask
flask-cors
sqlalchemy
psycopg2-binary
flask-socketio
eventlet